requirements.txt
python-dotenv
fastapi[all]
mcp
anthropic